Adding the input parameter Do_Pauli_blocking to all input files.
I added the input parameter Do_Pauli_blocking to all input files to avoid errors caused by not reading this parameter during execution. If this parameter is set to 1, it will multiply the inverse mean free path by 1 / (1 - f_eq); if set to 0, it will do nothing. In the tests that don't involve collisions, it will not have any effect.